webpack

+++ src/controllers/videoController.js
@@ -36,20 +36,28 @@ export const watch = async(req,res) => {
 
  export const getEdit = async(req,res) => {
    const { id } = req.params;
+   const { user: { _id }} = req.session;
    const video = await Video.findById(id);
    if (video === null) {
       return res.status(404).render("404",{pageTitle : "Video not found."});
    } 
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+   }
    return res.render("edit", {pageTitle : `Edit:${video.title} `, video});
 }; 
 
 export const postEdit = async(req,res) => {
    const { id } = req.params;
+   const { user: { _id }} = req.session;
    const { title, description, hashtags} = req.body;
    const video = await Video.exists({_id: id });
    if (!video) {
       return res.status(404).render("404",{pageTitle : "Video not found."});
    } 
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+   }
    await Video.findByIdAndUpdate(id, {
       title,
       description,
@@ -68,7 +76,7 @@ export const postUpload = async (req, res) => {
    const {path: fileUrl} = req.file;
    const { title, description, hashtags } = req.body;
    try {
-   await Video.create({ 
+   const newVideo = await Video.create({ 
       title,
       description,
       fileUrl,
@@ -76,6 +84,10 @@ export const postUpload = async (req, res) => {
       createdAt : Date.now(),
       hashtags : Video.formatHashtags(hashtags)
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
+
 
    /*
    const video = new Video({ 
@@ -105,6 +117,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req,res) => {
    const { id } = req.params;
+   const { user: { _id }} = req.session;
+   const video = await Video.findById(id);
+
+   if (!video) {
+      return res.status(404).render("404",{pageTitle : "Video not found."});
+   } 
+
+   if (String(video.owner) !== String(_id)) {
+      return res.status(403).redirect("/");
+   }
    await Video.findByIdAndDelete(id);
    return res.redirect("/");
     };
